+++ crypto/evp/build.info
@@ -4,6 +4,10 @@ $COMMON=digest.c evp_enc.c evp_lib.c evp_fetch.c cmeth_lib.c evp_utils.c \
         m_sigver.c pmeth_lib.c signature.c p_lib.c pmeth_gn.c exchange.c \
         pmeth_check.c evp_rand.c asymcipher.c kem.c
 
+# Diverse type specific ctrl functions.  They are kinda sorta legacy, kinda
+# sorta not.
+$COMMON=$COMMON dsa_ctrl.c
+
 SOURCE[../../libcrypto]=$COMMON\
         encode.c evp_key.c evp_cnf.c \
         e_des.c e_bf.c e_idea.c e_des3.c \
